requirements.txt
flask==3.1.2
requests==2.32.5
python-dotenv==1.2.1
gunicorn==21.2.0